Blog section refactored
