+++ Reminder.md
@@ -1,0 +1,9 @@
+# Reminders
+
+- Add SEO
+- Add Eslint configuration
+- GSAP or Explore Framer.js animations
+- enable smooth scrolling, scroll snap etc.
+- Image Optimization
+- Implement link functionality in blog
+- resume need to be up to dated
