+++ src/database/blogs.js
@@ -1,0 +1,5 @@
+import { blog1 } from './article';
+
+const blogs = [...blog1];
+
+export default blogs;
