+++ src/index.css
@@ -25,6 +25,7 @@ body {
   padding: 0;
   min-width: 320px;
   min-height: 100vh;
+  scroll-behavior: smooth;
 }
 
 @media (prefers-color-scheme: light) {
